+++ src/main/java/dev/willyelton/crystal_tools/levelable/skill/SkillData.java
@@ -4,12 +4,14 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import dev.willyelton.crystal_tools.utils.ListUtils;
 
+import java.util.Arrays;
 import java.util.List;
 
 public class SkillData {
     private final List<List<SkillDataNode>> nodes;
     // Flattened nodes, calculated lazily
     private List<SkillDataNode> flatNodes = null;
+    private int totalPoints;
 
     private SkillData(List<List<SkillDataNode>> nodes) {
         this.nodes = nodes;
@@ -21,6 +23,8 @@ public class SkillData {
         for (SkillDataNode node : nodes) {
             node.setPoints(points[node.getId()]);
         }
+
+        this.totalPoints = Arrays.stream(points).sum();
     }
 
     public List<List<SkillDataNode>> getAllNodesByTier() {
@@ -33,6 +37,15 @@ public class SkillData {
         }
 
         return flatNodes;
+    }
+
+    public int getTotalPoints() {
+        return totalPoints;
+    }
+
+    // TODO: This should be temporary, maybe include this whole class in the 1.21 redesign
+    public void addPoint() {
+        this.totalPoints++;
     }
 
     public static final Codec<SkillData> CODEC = RecordCodecBuilder.create(instance -> instance.group(